mcp-server/requirements.txt
mcp[cli]>=1.0.0
paho-mqtt>=2.0.0
sentence-transformers
numpy